Make dmdar take into account the size of missing data, not only their number
(cherry picked from commit 5b4da530f760f32cdc0c4e6dca8ceb8c706307bc)

+++ src/sched_policies/deque_modeling_policy_data_aware.c
@@ -152,7 +152,7 @@ static struct starpu_task *dmda_pop_ready_task(unsigned sched_ctx_id)
 #ifdef STARPU_VERBOSE
 		if (task->cl)
 		{
-			int non_ready = _starpu_count_non_ready_buffers(task, workerid);
+			size_t non_ready = _starpu_count_non_ready_buffers(task, workerid);
 			if (non_ready == 0)
 				dt->ready_task_cnt++;
 		}
